requirements.txt
bind==2.0.0
Django==2.2.4
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-phone-field==1.8.0
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
